backend/requirements.txt
requests>=2.32.3
beautifulsoup4>=4.12.3
playwright>=1.44.0
pdfplumber>=0.11.1
google-genai>=1.10.0
chromadb>=0.6.0
langchain-text-splitters>=0.2.2
sentence-transformers>=2.7.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.1
python-dotenv>=1.0.1
pydantic>=2.9.0
python-telegram-bot>=21.0.0
httpx>=0.27.0